+++ examples/aws/kinesis_put.py
@@ -1,0 +1,14 @@
+from yarl import URL
+
+from cratedb_toolkit.io.kinesis.adapter import KinesisAdapter
+from tests.io.test_processor import DYNAMODB_CDC_INSERT_NESTED, DYNAMODB_CDC_MODIFY_NESTED, wrap_kinesis
+
+
+def main():
+    ka = KinesisAdapter(URL("kinesis://LSIAQAAAAAAVNCBMPNSG:dummy@localhost:4566/cdc-stream?region=eu-central-1"))
+    ka.produce(wrap_kinesis(DYNAMODB_CDC_INSERT_NESTED))
+    ka.produce(wrap_kinesis(DYNAMODB_CDC_MODIFY_NESTED))
+
+
+if __name__ == "__main__":
+    main()